backend/requirements.txt
altair==5.5.0
altgraph==0.17.4
amqp==5.3.1
asgiref==3.10.0
attrs==25.3.0
babel==2.17.0
billiard==4.2.2
black==25.9.0
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.0
cairocffi==1.7.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cssselect2==0.8.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-redis==6.0.0
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-tenants==3.9.0
drf-spectacular==0.27.1
et_xmlfile==2.0.0
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
gunicorn==23.0.0
idna==3.10
isort==6.1.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
kombu==5.5.4
MarkupSafe==3.0.3
matplotlib==3.9.2
mypy_extensions==1.1.0
narwhals==2.5.0
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.5.0
plotly==6.3.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psycopg2-binary==2.9.11
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
pydyf==0.11.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyJWT==2.10.1
pyparsing==3.2.5
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
pywin32-ctypes==0.2.3
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
ruff==0.14.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.49.1
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tkcalendar==1.6.1
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.11.0
zopfli==0.4.0